add fake message class
